+++ orthanc_ext/event_dispatcher.py
@@ -1,14 +1,15 @@
 import json
 import logging
-
 from dataclasses import dataclass
 
-from orthanc_ext.logging_configurator import orthanc_logging
+from orthanc_ext.http_utilities import create_internal_client, get_rest_api_base_url, \
+    get_certificate
+from orthanc_ext.logging_configurator import python_logging
 from orthanc_ext.python_utilities import ensure_iterable, create_reverse_type_dict
-from orthanc_ext.requests_utilities import create_internal_requests_session, get_rest_api_base_url, get_certificate
 
 
-def register_event_handlers(event_handlers, orthanc_module, requests_session, logging_configuration=orthanc_logging):
+def register_event_handlers(
+        event_handlers, orthanc_module, requests_session, logging_configuration=python_logging):
     logging_configuration(orthanc_module)
 
     @dataclass
@@ -19,9 +20,10 @@ def register_event_handlers(event_handlers, orthanc_module, requests_session, lo
 
         def __str__(self):
             return (
-                f'ChangeEvent(change_type={event_types.get(self.change_type)}, '
-                f"resource_type={resource_types.get(self.resource_type)}, resource_id='{self.resource_id}')"
-            )
+                f'ChangeEvent('
+                f'change_type={event_types.get(self.change_type)}, '
+                f'resource_type={resource_types.get(self.resource_type)}, '
+                f"resource_id='{self.resource_id}')")
 
     def create_type_index(orthanc_type):
         return create_reverse_type_dict(orthanc_type)
@@ -47,6 +49,6 @@ def register_event_handlers(event_handlers, orthanc_module, requests_session, lo
 
 def create_session(orthanc):
     config = json.loads(orthanc.GetConfiguration())
-    return create_internal_requests_session(
-        get_rest_api_base_url(config), orthanc.GenerateRestApiAuthorizationToken(), get_certificate(config)
-    )
+    return create_internal_client(
+        get_rest_api_base_url(config), orthanc.GenerateRestApiAuthorizationToken(),
+        get_certificate(config))